pyproject.toml
[project]
name = "django-ecommerce-api"
version = "0.1.0"
description = "Comprehensive RESTful API for Django-based e-commerce applications with product management, user authentication, order processing, and payment integration features"
readme = "README.md"
requires-python = ">=3.6,<3.10"
dependencies = [
    "django>=3.2.25",
    "django-cors-headers>=3.10.1",
    "django-environ>=0.11.2",
    "djangorestframework>=3.15.1",
    "djangorestframework-simplejwt>=4.4.0",
    "djoser>=2.0.5",
    "pillow>=8.4.0",
    "stripe>=11.6.0",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=7.0.1",
    "pytest-django>=4.5.2",
    "ruff>=0.0.17",
]

[tool.ruff]
exclude = [
    "migrations",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

target-version = "py39"

line-length = 88
indent-width = 4


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"


[tool.ruff.lint]
select = ["F", "E", "I", "E4", "E7", "E9"]
fixable = ["ALL"]


[tool.pyright]
include = ["ecommerce_django"]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "src/experimental",
    "src/typestubs",
]
reportMissingImports = "error"
reportMissingTypeStubs = false
reportIncompatibleVariableOverride = "none"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "ecommerce_django.settings"
pythonpath = [
  "."
]
testpaths = ["ecommerce_django/tests"]
python_files = "test_*.py"